requirements.txt
Django==1.7.1
django-braces==1.4.0
django-crispy-forms==1.4.0
django-phonenumber-field==0.6
gunicorn==19.1.1
httplib2==0.9
path.py==7.0
phonenumbers==7.0.0
psycopg2==2.5.4
pytz==2014.9
requests==2.4.3
twilio==3.6.9
vobject==0.6.6
MySQL-python==1.2.5
django-selectize==1.2
python-social-auth==0.2.13
django-extra-views
Markdown==2.3.1
markdown2
premailer
djrill
django-celery
whitenoise
dj-database-url